mail, tests: Rearrange test file and smtp instance
smtp/smtp.py is bad way to store implementation of protocol. There
might be up to three files in such directory. But we'll implement
more than three protocols in the future.

The major reason of rearranging dummy_mail test case is that we're
not developing too many protocol recently. No need to consider
distant future.